Add small fixes
- avoid proxy use in "chekc for updates" for spossible speedup
- ignore optional windows updates

+++ src/VM Management Tool/Configuration/ConfigurationManager.cs
@@ -410,7 +410,7 @@ namespace VMManagementTool
                     "log",
                     new Dictionary<string,object>
                     {
-                        { "level", 2}
+                        { "level", (long)2}
                     }
                 },
             };
